+++ core/src/main/java/org/seedstack/seed/core/internal/init/NoOpLogManager.java
@@ -20,4 +20,9 @@ class NoOpLogManager implements LogManager {
     public void close() {
         // noop
     }
+
+    @Override
+    public void refresh(LoggingConfig loggingConfig) {
+        // noop
+    }
 }